Refactor approach so it works better
We can't really detect the top-level programs first, because by that point, Roslyn already knows about the multiple files containing them and will show error squiglies in the IDE which are quite disconcerting when things actually do compile fine later on.

So instead, we collect all top-level C# files (that is, not within a subfolder) and assume those will be the top-level programs. We add them all as launch profiles, and then before compile we remove all the Compile items which don't match the active debug profile selected. This works much better, even if intellisense is only on for the file which is selected in the dropdown, even if it's not the active one in the editor. Making both match will likely require IDE extensibility, so it's out of scope (for now?).