Make exprience smoother with intellisense
Adding/removing items from implicit Compile default items was causing intellisense to not refresh appropriately.

So instead we adopt a more declarative approach, modifying the default globing to not include the root .cs files and only globing from subfolders (*/**/*.cs instead of simply **/*.cs), and we add all root .cs files as content instead. Then we declaratively only add the Compile for the 'ActiveDebugProfile' name/file if it exists. This would almost always match and be the right selection.

For the rare cases where it doesn't exist or is not set yet to a value (i.e. prior to the first run after installing SmallSharp), we determine if the .user project exists, and create or update it accordingly from the first .cs we find in the project.

That's the only case where we'll manually add the <Compile /> item, and on subsequent builds the target condition will not be satisfied and the .props inclusion will kick in instead, causing the intellisense to refresh properly.

The only drawback of this approach is that changing the startup file selection will close the current Compile file if it was the startup one, since it changes item type completely and CPS doesn't seem to be able to deal with that. Also, since we're not extending the IDE, we're not able to open the subsequently selected file either.