Add more resiliency when communicating with DTE
Since this is a cross-process operation that may involve attempting to acquire the UI thread while the target VS is actually busy, it may actually fail with a COM exception. In that case, default to not throwing, since that destabilizes the MSBuild process and causes us to not recover the active tracking once the UI thread lock is gone.