Populate launchSettings with candidate top-level compile files
We used to do this via a C# source generator, but stopped doing it because it interferred with the compilation and sometimes broke intellisense.

Now we have a more reliable auto-selection mechanism based on detecting the active document. In some rare cases, however, the IDE detection stops working and it's impossible to select the actual target we want to run.

This brings back the launchSettings.json generation, but based entirely on MSBuild instead, so we don't interfere at all with any of the built-in stuff that's already working. We should also run this right after we select the startup file, so it's basically a complement, not a replacement.

Fixes #40