Add support for #:sdk directive and no restore failures
The SDK mode allows us to support the #:sdk directive such that our SDK imports the generated .sdk.props and .sdk.targets files. It can also set the required `ImportProjectExtension*` properties for automatically importing the targets during restore (if present in the base intermediate output path).

Even with the added properties, the situation on initial clone/run of a project using SmallSharp was suboptimal:
- whether in SDK mode or package mode, we'd need a second restore after the initial one and first EmitTargets run. VS does this automatically since it would detect the addition of new extension targets (due to the ImportProjectExtension* props) and restore again as needed. But it wouldn't work from the CLI
- this meant a dotnet build would fail consistently

So we have a dual mechanism that makes this seamless:
- In SDK mode we can optimize things by just parsing the startup/active file and injecting the package references before the restore graph is generated
- In package mode, we re-run restore and temporarily redirect the assets file to a different dynamic one to refresh the resolved assets.

Fixes #143