Ensure when running from VS, output supports UTF-8
The console started by VS (even if it's the terminal one) by default does not seem to be set up to support UTF-8. To be on the conservative side, we just set up the encoding in DEBUG builds so things Just Work from VS by default.

This problem doesn't exist when running from a regular terminal.